LayoutTests/fast/js/Promise-simple.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description('Test Promise.');

window.jsTestIsAsync = true;

var resolver;

var firstPromise = new Promise(function(newResolver) {
  window.thisInInit = this;
  resolver = newResolver;
});

var secondPromise = firstPromise.then(function(result) {
  window.thisInFulfillCallback = this;
  shouldBeTrue('thisInFulfillCallback === secondPromise');
  window.result = result;
  shouldBeEqualToString('result', 'hello');
  finishJSTest();
});

shouldBeTrue('thisInInit === firstPromise');

resolver.fulfill('hello');

</script>
<script src="resources/js-test-post.js"></script>
</body>
</html>
