LayoutTests/fast/js/Promise-simple-expected.txt

Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS thisInInit === firstPromise is true
PASS thisInFulfillCallback === secondPromise is true
PASS result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
